decimer-smiles/azure/requirements.txt
decimer
psutil
psycopg2-binary
azure-storage-blob
